requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==2.2.2
Flask==1.0.2
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.18
PyNaCl==1.2.1
python-dotenv==0.8.2
six==1.11.0
SQLAlchemy==1.2.8
SQLAlchemy-Utils==0.33.3
Werkzeug==0.14.1
